fix: 修复po，修改原始数据

+++ src/main/java/com/example/hotel/po/CreditChange.java
@@ -17,6 +17,6 @@ public class CreditChange {
     private String reason;
     private Double changeNum;
     private Integer orderId;
-    private Integer credit;
+    private Double credit;
     private Timestamp time;
 }
